// kitchen-planner-fe-main/src/app/api/api-kms-planner-masterdata/model/category.ts
/**
 * KMS - Kitchen Management System API
 *
 * Contact: anas-ur-rasheed.abid@tu-ilmenau.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Category { 
    _id?: string;
    name?: string;
    description?: string;
    isActive?: boolean;
    createdAt?: string;
}